+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -1,13 +1,13 @@
 package com.kodilla.basic_assertion;
 
 public class Calculator {
-    public int sum(int a, int b) {
+    public double sum(double a, double b) {
         return a+b;
     }
-    public int sub(int a, int b) {
+    public double sub(double a, double b) {
         return a-b;
     }
-    public int square(int a) {
+    public double square(double a) {
         return a*a;
     }
 }
